New Method1
New Method1

+++ alvr/client_core/cpp/bindings.h
@@ -44,7 +44,7 @@ extern "C" void renderLobbyNative(const FfiViewInput eyeInputs[2]);
 extern "C" void renderStreamNative(void *streamHardwareBuffer,
                                    const unsigned int swapchainIndices[2]);
 extern "C" void updategussionflg( bool flag , int strategynum, float roisize);
-extern "C" void updategazecenter( unsigned long long targetTimestampNs ,
+extern "C" void updategazecenter( __uint128_t targetTimestampNs ,
                                 float headx, float heady, float gazex, float gazey ,
                                 float lx,float ly ,float rx ,float ry
                                 );
